Move the Phoenix-lifted functions into their own module.
